+++ packages/core-gatsby-ui/src/design-system-docs/components/navigation.page.tsx
@@ -1,5 +1,6 @@
+/* eslint-disable react/jsx-no-undef */
 import { HEADING } from '@newrade/core-design-system';
-import { BoxV2, Heading, NavBar, Stack, useTreatTheme } from '@newrade/core-react-ui';
+import { BoxV2, Heading, Label, NavBar, Stack, useTreatTheme } from '@newrade/core-react-ui';
 import React from 'react';
 import { useStyles } from 'react-treat';
 import { DesignSystemPageProps, DesignSystemPageTemplate } from '../../templates/design-system-page.template';
@@ -21,12 +22,19 @@ const PageComponent: React.FC<DesignSystemPageProps> = (props) => {
 
       <Stack gap={[cssTheme.sizing.var.x5]}>
         <Heading variant={HEADING.h2}>NavBar</Heading>
-
         <BoxV2 padding={[cssTheme.sizing.var.x3, 0]}>
           <NavBar
             style={{ position: 'relative' }}
             MobileSvgLogo={<LogoPlaceholder height={'100%'} />}
             DesktopSvgLogo={<LogoPlaceholder height={'100%'} />}
+            MenuLinks={
+              <>
+                <Label>Services</Label>
+                <Label>Technologies</Label>
+                <Label>Agence</Label>
+                <Label>Contact</Label>
+              </>
+            }
           />
         </BoxV2>
       </Stack>