+++ packages/core-webpack-config/src/babel-preset.conf.ts
@@ -3,6 +3,17 @@
  */
 export const babelPresetConf: ([string, object] | [string])[] = [
   /**
+   * @see https://babeljs.io/docs/en/babel-preset-typescript
+   */
+  [
+    '@babel/preset-typescript',
+    {
+      isTSX: true, // defaults to false
+      jsxPragma: `jsx`, // defaults to "React"
+      allExtensions: true, // defaults to false
+    },
+  ],
+  /**
    * A Babel preset that compiles ES2015+ down to ES5 by automatically determining the Babel plugins
    * and polyfills you need based on your targeted browser or runtime environments.
    *
@@ -15,9 +26,11 @@ export const babelPresetConf: ([string, object] | [string])[] = [
       // see .browserslist -> for target which is a query to select browsers (ex: last 2 versions, > 5%, safari tp) using browserslist.
       targets: ['chrome >= 85', 'firefox >= 82', 'ios_saf >=12', 'safari >=12', 'not IE 11', 'node >= 14'],
       useBuiltIns: 'usage', // Adds specific imports for polyfills when they are used in each file.
-      corejs: 3,
-      shippedProposals: false, // Toggles enabling support for builtin/feature proposals that have shipped in browsers.
+      corejs: '3.8',
+      shippedProposals: true, // Toggles enabling support for builtin/feature proposals that have shipped in browsers.
       modules: false, // don't transform modules (webpack will do it)
+      loose: true,
+      ignoreBrowserslistConfig: true,
     },
   ],
   ['@babel/preset-react'],