+++ packages/mir-website/src/components/navigation/nav-bar.tsx
@@ -2,7 +2,7 @@ import { Link as GatsbyLink } from 'gatsby';
 import React from 'react';
 import { useStyles } from 'react-treat';
 import Icon from '../../icons/IconBars.svg';
-import { PAGE_NAME } from '../../templates/page.template';
+import { PAGE_NAME } from '../../templates/contentful-page.template';
 import { Button } from '../ui/button';
 import { Link } from '../ui/link';
 import * as stylesRef from './nav-bar.treat';