+++ packages/core-design-system-old/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/newrade/newrade-core/tree/master/packages/core-design-system/compare/@newrade/core-design-system-old@1.4.0...@newrade/core-design-system-old@1.4.1) (2021-02-19)
+
+**Note:** Version bump only for package @newrade/core-design-system-old
+
 # [1.4.0](https://github.com/newrade/newrade-core/tree/master/packages/core-design-system/compare/@newrade/core-design-system-old@1.3.5...@newrade/core-design-system-old@1.4.0) (2021-02-11)
 
 ### Features